Made sure Pythonqt preparation script checks out svn trunk for building with Qt 4.8 on precise

+++ HOPSAN++/HopsanGUI/Dependencies/unpackPatchAndBuildPythonQt.sh
@@ -5,7 +5,7 @@
 # Author: Peter Nordin peter.nordin@liu.se
 # Date:   2012-03-29
 
-pythonqtname="PythonQt2.0.1"
+
 pyversion="2.7"
 basepwd=`pwd`
 
@@ -13,19 +13,45 @@ if [ "$2" != "" ]; then
   pyversion="$2"  
 fi
 
-# Abort if dir already exist. When running release build script we dont want to build twice
-if [ -d $pythonqtname ]; then
-  echo Directory $pythonqtname already exist. Remove it if you want build using this script.
-  exit 0
+ubuntuversion=$(echo `lsb_release -rs` | sed 's|\.||')
+
+if [ $ubuntuversion -lt "1204" ]; then
+
+  pythonqtname="PythonQt2.0.1"
+
+  # Abort if dir already exist. When running release build script we dont want to build twice
+  if [ -d $pythonqtname ]; then
+    echo Directory $pythonqtname already exist. Remove it if you want build using this script.
+    exit 0
+  fi
+
+  rm -rf $pythonqtname
+  unzip -q $pythonqtname.zip
+    
+  echo "Applying Hopsan fixes to code"
+  # Apply patch to remove qt all extensions and cocoa thing
+  patch -p0 < PythonQt201_ubuntu.patch
+
+else
+
+  pythonqtname="PythonQtr209"
+
+  # Abort if dir already exist. When running release build script we dont want to build twice
+  if [ -d $pythonqtname ]; then
+    echo Directory $pythonqtname already exist. Remove it if you want build using this script.
+    exit 0
+  fi
+
+  svn co -r 209 https://pythonqt.svn.sourceforge.net/svnroot/pythonqt/trunk $pythonqtname
+
+  echo "Applying Hopsan fixes to code"
+  # Apply patch to remove qt all extensions and cocoa thing
+  patch -p0 < PythonQtr209_ubuntu.patch
+
+  #Remove tests and examples, at least test will not build
+  sed "s|tests examples||" -i PythonQt.pro
+
 fi
-
-rm -rf $pythonqtname
-unzip -q $pythonqtname.zip
-
-  
-echo "Applying Hopsan fixes to code"
-# Apply patch to remove qt all extensions and cocoa thing
-patch -p0 < PythonQt201_ubuntu.patch
 
 cd $pythonqtname
